Implement a stock_picker method
Add the stock_picker method which creates a bidimensional array
with all possible buy and sell operations, then finds the highest
possible value, find the buy and sell day of this value and then
return both in an array.

This method follows every project's requirement, it works for any
array with length bigger than 1 and in the case of negative trades
it chooses the least worst case.